refactor: delete conslole.log (#34)
delete duplicate redux hooks
