+++ src/components/ui/ParticlesBackground/ParticlesBackground.tsx
@@ -8,7 +8,7 @@ export const ParticlesBackground = () => {
   }, []);
 
   const particlesLoaded = useCallback(async (container: Container | undefined) => {
-    await console.log(container);
+    await container;
   }, []);
 
   return (
